Aggiunta di carrello Bean e model
