+++ OhMyBag/src/it/ohmybag/model/Carrello.java
@@ -1,0 +1,5 @@
+package it.ohmybag.model;
+
+public class Carrello {
+
+}
